Renamed AllVHDLSources to Sanity. Improved glob in Sanity check. Bound requirement for pyVHDLModel to Git branch.
