+++ testsuite/pyunit/dom/Sanity.py
@@ -7,9 +7,9 @@
 # |_|    |___/
 # =============================================================================
 # Authors:
-#   Patrick Lehmann
+#   Unai Martinez-Corral
 #
-# Testsuite:        Check libghdl IIR translation to DOM.
+# Testsuite:        Parse files from sanity checks
 #
 # License:
 # ============================================================================
@@ -30,3 +30,24 @@
 #
 # SPDX-License-Identifier: GPL-2.0-or-later
 # ============================================================================
+from sys import executable
+from subprocess import check_call, STDOUT
+from pathlib  import Path
+from pytest import mark
+
+if __name__ == "__main__":
+    print("ERROR: you called a testcase declaration file as an executable module.")
+    print("Use: 'python -m unitest <testcase module>'")
+    exit(1)
+
+_TESTSUITE_ROOT = Path(__file__).parent.parent.parent.resolve()
+_GHDL_ROOT = _TESTSUITE_ROOT.parent
+
+@mark.xfail
+@mark.parametrize("file", [str(f) for f in _TESTSUITE_ROOT.glob("sanity/**/*.vhdl")])
+def test_AllVHDLSources(file):
+    check_call([
+        executable,
+        _GHDL_ROOT / "pyGHDL/cli/DOM.py",
+        file
+    ], stderr=STDOUT)
